refactor(coordinator): split api/consumer.go into focused files
Same-package modularization of the consumer god-file (no behavior change):
  - httputil.go: writeJSON + the OpenAI-compatible errorResponse envelope
  - responses_translate.go: Responses-API request lowering (input/tools/
    tool_choice/text.format -> chat-completions shape)
  - models_endpoints.go: GET /v1/models and /v1/models/{id} + alias listing

consumer.go 4884 -> 3794 LOC. Full api go test ./api/... stays green.
