+++ coordinator/api/consumer.go
@@ -1008,235 +1008,6 @@ func copyJSONMap(in map[string]any) map[string]any {
 	return out
 }
 
-func responsesContentText(content any) string {
-	switch c := content.(type) {
-	case nil:
-		return ""
-	case string:
-		return c
-	case []any:
-		parts := make([]string, 0, len(c))
-		for _, part := range c {
-			switch p := part.(type) {
-			case string:
-				if p != "" {
-					parts = append(parts, p)
-				}
-			case map[string]any:
-				if text, _ := p["text"].(string); text != "" {
-					parts = append(parts, text)
-					continue
-				}
-				if p["type"] == "input_image" || p["type"] == "input_file" {
-					// Text models cannot consume binary Responses parts yet.
-					// Preserve the turn shape without leaking URLs or blobs.
-					parts = append(parts, fmt.Sprintf("[%s omitted]", p["type"]))
-				}
-			}
-		}
-		return strings.Join(parts, "\n")
-	default:
-		b, err := json.Marshal(c)
-		if err != nil {
-			return fmt.Sprint(c)
-		}
-		return string(b)
-	}
-}
-
-func responsesInputToChatMessages(input any) ([]map[string]any, error) {
-	switch v := input.(type) {
-	case string:
-		return []map[string]any{{"role": "user", "content": v}}, nil
-	case []any:
-		messages := make([]map[string]any, 0, len(v))
-		for _, raw := range v {
-			item, ok := raw.(map[string]any)
-			if !ok {
-				continue
-			}
-			if typ, _ := item["type"].(string); typ != "" {
-				switch typ {
-				case "message":
-					role, _ := item["role"].(string)
-					if role == "" {
-						role = "user"
-					}
-					if role == "developer" {
-						role = "system"
-					}
-					messages = append(messages, map[string]any{
-						"role":    role,
-						"content": responsesContentText(item["content"]),
-					})
-				case "function_call":
-					callID, _ := item["call_id"].(string)
-					if callID == "" {
-						callID, _ = item["id"].(string)
-					}
-					name, _ := item["name"].(string)
-					args, _ := item["arguments"].(string)
-					messages = append(messages, map[string]any{
-						"role":    "assistant",
-						"content": "",
-						"tool_calls": []map[string]any{{
-							"id":   callID,
-							"type": "function",
-							"function": map[string]any{
-								"name":      name,
-								"arguments": args,
-							},
-						}},
-					})
-				case "function_call_output":
-					callID, _ := item["call_id"].(string)
-					messages = append(messages, map[string]any{
-						"role":         "tool",
-						"tool_call_id": callID,
-						"content":      responsesContentText(item["output"]),
-					})
-				case "reasoning":
-					// Reasoning items are model-side metadata, not prompt text.
-					continue
-				default:
-					return nil, fmt.Errorf("unsupported Responses input item type %q", typ)
-				}
-				continue
-			}
-
-			role, _ := item["role"].(string)
-			if role == "" {
-				continue
-			}
-			if role == "developer" {
-				role = "system"
-			}
-			messages = append(messages, map[string]any{
-				"role":    role,
-				"content": responsesContentText(item["content"]),
-			})
-		}
-		if len(messages) == 0 {
-			return nil, fmt.Errorf("Responses input did not contain any chat-compatible messages")
-		}
-		return messages, nil
-	default:
-		return nil, fmt.Errorf("Responses input must be a string or array")
-	}
-}
-
-func responsesToolsToChatTools(raw any) ([]any, error) {
-	tools, ok := raw.([]any)
-	if !ok || len(tools) == 0 {
-		return nil, nil
-	}
-	out := make([]any, 0, len(tools))
-	for _, rawTool := range tools {
-		tool, ok := rawTool.(map[string]any)
-		if !ok {
-			continue
-		}
-		typ, _ := tool["type"].(string)
-		if typ == "" || typ == "function" {
-			name, _ := tool["name"].(string)
-			if name == "" {
-				if fn, _ := tool["function"].(map[string]any); fn != nil {
-					out = append(out, tool)
-					continue
-				}
-				return nil, fmt.Errorf("function tool is missing name")
-			}
-			fn := map[string]any{"name": name}
-			if description, ok := tool["description"].(string); ok {
-				fn["description"] = description
-			}
-			if parameters, ok := tool["parameters"]; ok {
-				fn["parameters"] = parameters
-			}
-			out = append(out, map[string]any{
-				"type":     "function",
-				"function": fn,
-			})
-			continue
-		}
-		return nil, fmt.Errorf("unsupported Responses tool type %q", typ)
-	}
-	return out, nil
-}
-
-func responsesToolChoiceToChat(raw any) (any, error) {
-	choice, ok := raw.(map[string]any)
-	if !ok {
-		return raw, nil
-	}
-	typ, _ := choice["type"].(string)
-	if typ == "function" {
-		name, _ := choice["name"].(string)
-		if name == "" {
-			return nil, fmt.Errorf("function tool_choice is missing name")
-		}
-		return map[string]any{
-			"type": "function",
-			"function": map[string]any{
-				"name": name,
-			},
-		}, nil
-	}
-	return raw, nil
-}
-
-func responsesTextFormatToChatResponseFormat(raw any) any {
-	text, ok := raw.(map[string]any)
-	if !ok {
-		return nil
-	}
-	format, ok := text["format"].(map[string]any)
-	if !ok {
-		return nil
-	}
-	switch format["type"] {
-	case "json_object":
-		return map[string]any{"type": "json_object"}
-	case "json_schema":
-		return map[string]any{
-			"type":        "json_schema",
-			"json_schema": format,
-		}
-	}
-	return nil
-}
-
-func responsesRequestToChatCompletions(parsed map[string]any) (map[string]any, error) {
-	messages, err := responsesInputToChatMessages(parsed["input"])
-	if err != nil {
-		return nil, err
-	}
-
-	out := copyJSONMap(parsed)
-	delete(out, "input")
-	delete(out, "endpoint")
-	delete(out, "max_output_tokens")
-	delete(out, "text")
-	out["messages"] = messages
-	if maxTokens := explicitMaxTokens(parsed); maxTokens > 0 {
-		out["max_tokens"] = maxTokens
-	}
-	if tools, err := responsesToolsToChatTools(parsed["tools"]); err != nil {
-		return nil, err
-	} else if len(tools) > 0 {
-		out["tools"] = tools
-	}
-	if choice, err := responsesToolChoiceToChat(parsed["tool_choice"]); err != nil {
-		return nil, err
-	} else if choice != nil {
-		out["tool_choice"] = choice
-	}
-	if responseFormat := responsesTextFormatToChatResponseFormat(parsed["text"]); responseFormat != nil {
-		out["response_format"] = responseFormat
-	}
-	return out, nil
-}
-
 // handleChatCompletions handles POST /v1/chat/completions.
 //
 // This is the main inference endpoint. It validates the request, finds an
@@ -3005,258 +2776,6 @@ func buildNonStreamingResponse(requestID, model string, msg extractedMessage, us
 // from the catalog can't appear in the listing anyway, and adding it would
 // pollute the hidden set (e.g. an alias pointing at a not-yet-registered build).
 // Empty ids are ignored.
-func hideAliasBuild(hidden map[string]struct{}, catalogByID map[string]store.SupportedModel, buildID string) {
-	if buildID == "" {
-		return
-	}
-	if _, inCatalog := catalogByID[buildID]; inCatalog {
-		hidden[buildID] = struct{}{}
-	}
-}
-
-// aliasModelEntries builds the consumer-facing /v1/models entries for active
-// public aliases and returns the set of underlying build ids those aliases
-// cover (so the caller can hide them from the default listing). The hidden set
-// covers EVERY build an alias references — desired, previous, and the retired
-// lineage — so a concrete quant build never appears as its own entry once it is
-// behind an alias. Each alias entry derives its metadata from its primary build
-// — the desired build, or the previous build if the desired one isn't in the
-// catalog yet — and aggregates live capacity across the desired and previous
-// builds so the alias's routable/warm counts reflect every quant currently
-// serving it (retired builds are hide-only and never contribute capacity).
-func (s *Server) aliasModelEntries(
-	capByModel map[string]*registry.ModelCapacity,
-	catalogByID map[string]store.SupportedModel,
-	registryByID map[string]store.ModelRegistryEntry,
-) ([]types.ModelEntry, map[string]struct{}) {
-	hidden := make(map[string]struct{})
-	aliases, err := s.store.ListModelAliases()
-	if err != nil {
-		s.logger.Error("model registry: failed to list aliases", "error", err)
-		return nil, hidden
-	}
-
-	entries := make([]types.ModelEntry, 0, len(aliases))
-	for _, a := range aliases {
-		if !a.Active || a.DesiredBuild == "" {
-			continue
-		}
-		// A consumer must only ever see the alias, never a concrete build behind
-		// it. Hide EVERY build this alias references — desired, previous, AND the
-		// retired lineage — from the standalone listing, even if the alias itself
-		// isn't advertisable right now. (Capacity below aggregates only the
-		// routable desired/previous members; retired builds are hide-only.)
-		hideAliasBuild(hidden, catalogByID, a.DesiredBuild)
-		hideAliasBuild(hidden, catalogByID, a.PreviousBuild)
-		for _, b := range a.RetiredBuilds {
-			hideAliasBuild(hidden, catalogByID, b)
-		}
-		// Primary build = the desired build when it's in the catalog, else the
-		// previous build (so the alias keeps a real entry while the desired build
-		// is mid-registration). An alias whose builds are all out of catalog
-		// resolves to nothing and must not be advertised (it would 503).
-		members := make([]string, 0, 2)
-		desiredInCatalog := false
-		if _, ok := catalogByID[a.DesiredBuild]; ok {
-			members = append(members, a.DesiredBuild)
-			desiredInCatalog = true
-		}
-		previousInCatalog := false
-		if a.PreviousBuild != "" {
-			if _, ok := catalogByID[a.PreviousBuild]; ok {
-				members = append(members, a.PreviousBuild)
-				previousInCatalog = true
-			}
-		}
-		var primary string
-		switch {
-		case desiredInCatalog:
-			primary = a.DesiredBuild
-		case previousInCatalog:
-			primary = a.PreviousBuild
-		default:
-			// No in-catalog build backs this alias — don't advertise it.
-			continue
-		}
-
-		routable, warm := 0, 0
-		canAccept := false
-		for _, b := range members {
-			if cap, ok := capByModel[b]; ok {
-				routable += cap.RoutableProviders
-				warm += cap.WarmProviders
-				canAccept = canAccept || cap.CanAccept
-			}
-		}
-
-		cm := catalogByID[primary]
-		reg, hasReg := registryByID[primary]
-		displayName := a.DisplayName
-		if displayName == "" {
-			displayName = cm.DisplayName
-		}
-		metadata := types.ModelMetadata{
-			ModelType:         cm.ModelType,
-			Quantization:      "", // an alias spans quants; omit the per-build quant
-			DisplayName:       displayName,
-			RoutableProviders: routable,
-			WarmProviders:     warm,
-			CanAccept:         canAccept,
-		}
-		entry := types.ModelEntry{
-			ID:       a.AliasID,
-			Object:   "model",
-			OwnedBy:  "eigeninference",
-			Name:     displayName,
-			Metadata: metadata,
-		}
-		// Pricing / context / features come from the primary build's registry
-		// entry. Quantization is intentionally left blank on the alias.
-		primaryQuant := ""
-		if hasReg {
-			primaryQuant = reg.Quantization
-		}
-		s.openRouterModelFieldsFor(primary, primaryQuant, reg, hasReg).applyToModelEntry(&entry)
-		entry.Quantization = ""
-		var caps []string
-		if hasReg {
-			caps = reg.Capabilities
-		}
-		entry.InputModalities, entry.OutputModalities = deriveModalities(cm.ModelType, caps)
-		entries = append(entries, entry)
-	}
-	return entries, hidden
-}
-
-// listModelEntries assembles the consumer-facing model entries shared by
-// GET /v1/models and GET /v1/models/{id}. includeBuilds also lists the raw
-// quant builds hidden behind public aliases (ops/debug).
-func (s *Server) listModelEntries(includeBuilds bool) ([]types.ModelEntry, error) {
-	models := s.registry.ListModels()
-
-	// Build a lookup of capacity data keyed by model ID.
-	capacities := s.registry.ModelCapacitySnapshot()
-	capByModel := make(map[string]*registry.ModelCapacity, len(capacities))
-	for i := range capacities {
-		capByModel[capacities[i].ModelID] = &capacities[i]
-	}
-
-	// Filter to only show models from the active catalog, and capture the richer
-	// registry entries used to populate the OpenRouter provider fields. These
-	// lookups are shared with the dedicated /v1/models/openrouter feed.
-	catalogByID, registryByID, err := s.activeCatalogLookups()
-	if err != nil {
-		return nil, err
-	}
-
-	// Public aliases are the consumer-facing model names; their underlying
-	// quant builds are hidden by default so consumers never see the quant.
-	aliasEntries, hiddenBuilds := s.aliasModelEntries(capByModel, catalogByID, registryByID)
-
-	data := make([]types.ModelEntry, 0, len(models)+len(aliasEntries))
-	data = append(data, aliasEntries...)
-	for _, m := range models {
-		cm, inCatalog := catalogByID[m.ID]
-		if len(catalogByID) > 0 && !inCatalog {
-			continue
-		}
-		if _, hidden := hiddenBuilds[m.ID]; hidden && !includeBuilds {
-			continue
-		}
-		metadata := types.ModelMetadata{
-			ModelType:         m.ModelType,
-			Quantization:      m.Quantization,
-			ProviderCount:     m.Providers,
-			AttestedProviders: m.AttestedProviders,
-			TrustLevel:        string(m.TrustLevel),
-		}
-		// Add capacity fields from live snapshot.
-		if cap, ok := capByModel[m.ID]; ok {
-			metadata.RoutableProviders = cap.RoutableProviders
-			metadata.WarmProviders = cap.WarmProviders
-			metadata.CanAccept = cap.CanAccept
-		} else {
-			metadata.RoutableProviders = 0
-			metadata.WarmProviders = 0
-			metadata.CanAccept = false
-		}
-		if m.Attestation != nil {
-			metadata.Attestation = &types.ModelAttestation{
-				SecureEnclave: m.Attestation.SecureEnclave,
-				SIPEnabled:    m.Attestation.SIPEnabled,
-				SecureBoot:    m.Attestation.SecureBoot,
-			}
-		}
-		if inCatalog && cm.DisplayName != "" {
-			metadata.DisplayName = cm.DisplayName
-		}
-
-		entry := types.ModelEntry{
-			ID:            m.ID,
-			Object:        "model",
-			Created:       0,
-			OwnedBy:       "eigeninference",
-			Name:          metadata.DisplayName,
-			HuggingFaceID: m.ID, // model IDs are HuggingFace paths
-			Metadata:      metadata,
-		}
-
-		// OpenRouter provider fields (quantization, per-token pricing, sampling
-		// params, and registry-sourced metadata), shared with the dedicated
-		// /v1/models/openrouter feed.
-		reg, hasReg := registryByID[m.ID]
-		s.openRouterModelFieldsFor(m.ID, m.Quantization, reg, hasReg).applyToModelEntry(&entry)
-
-		// Modalities are derived from the model's capabilities (text by default).
-		var caps []string
-		if hasReg {
-			caps = reg.Capabilities
-		}
-		entry.InputModalities, entry.OutputModalities = deriveModalities(m.ModelType, caps)
-
-		data = append(data, entry)
-	}
-
-	return data, nil
-}
-
-func (s *Server) handleListModels(w http.ResponseWriter, r *http.Request) {
-	// Pass ?include_builds=1 (ops/debug) to also list the raw quant builds.
-	data, err := s.listModelEntries(r.URL.Query().Get("include_builds") == "1")
-	if err != nil {
-		s.logger.Error("model registry: failed to list active models", "error", err)
-		writeJSON(w, http.StatusInternalServerError, errorResponse("internal_error", "failed to list models"))
-		return
-	}
-
-	writeJSON(w, http.StatusOK, types.ModelListResponse{
-		Object: "list",
-		Data:   data,
-	})
-}
-
-// handleGetModel handles GET /v1/models/{id...} — the OpenAI "retrieve model"
-// endpoint. Model IDs may contain slashes (HuggingFace paths), hence the
-// wildcard path segment. Hidden quant builds are retrievable by their exact
-// id, matching the behavior of requesting one for inference.
-func (s *Server) handleGetModel(w http.ResponseWriter, r *http.Request) {
-	id := r.PathValue("id")
-	data, err := s.listModelEntries(true)
-	if err != nil {
-		s.logger.Error("model registry: failed to list active models", "error", err)
-		writeJSON(w, http.StatusInternalServerError, errorResponse("internal_error", "failed to list models"))
-		return
-	}
-	for _, entry := range data {
-		if entry.ID == id {
-			writeJSON(w, http.StatusOK, entry)
-			return
-		}
-	}
-	writeJSON(w, http.StatusNotFound, errorResponse("model_not_found",
-		fmt.Sprintf("model %q not found", id), withParam("model")))
-}
-
 // ── Multi-key management handlers (GET/POST/PATCH/DELETE /v1/keys) ────
 
 // createAPIKeyRequest is the POST /v1/keys (and rotate inherit) body. Money is
@@ -3469,14 +2988,6 @@ func (s *Server) handleProviderEarnings(w http.ResponseWriter, r *http.Request) 
 }
 
 // --- helpers ---
-
-// writeJSON serializes v as JSON and writes it to the response with the
-// given HTTP status code. Sets Content-Type to application/json.
-func writeJSON(w http.ResponseWriter, status int, v any) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	json.NewEncoder(w).Encode(v)
-}
 
 // handleCompletions handles POST /v1/completions.
 // Proxies OpenAI-compatible text completions to the selected provider over the
@@ -4280,37 +3791,3 @@ func (s *Server) handleGenericInference(w http.ResponseWriter, r *http.Request, 
 		s.handleNonStreamingResponseWithFirstChunk(w, r, pr, firstChunks)
 	}
 }
-
-// errorDetailOpt carries optional fields for OpenAI-compatible error responses.
-type errorDetailOpt struct {
-	param string // e.g. "model", "max_tokens"
-	code  string // e.g. "model_not_found", "insufficient_quota"
-}
-
-// errorResponse builds a standard OpenAI-compatible error response body.
-// By default, code is inferred from errType. Callers can override code or
-// set param via withParam / withCode helpers.
-func errorResponse(errType, message string, opts ...errorDetailOpt) map[string]any {
-	detail := map[string]any{
-		"type":    errType,
-		"message": message,
-		"code":    errType, // default: code mirrors type
-	}
-	for _, o := range opts {
-		if o.param != "" {
-			detail["param"] = o.param
-		}
-		if o.code != "" {
-			detail["code"] = o.code
-		}
-	}
-	return map[string]any{
-		"error": detail,
-	}
-}
-
-// withParam returns an option that sets the "param" field on an error response.
-func withParam(p string) errorDetailOpt { return errorDetailOpt{param: p} }
-
-// withCode returns an option that overrides the "code" field on an error response.
-func withCode(c string) errorDetailOpt { return errorDetailOpt{code: c} }
